Set up expanded state for StudentCard
This commit adds an `expanded` boolean state to the `<StudentCard />`
component. When this state is true, the card will display all of the
student's grades. When false, grades will not be shown.

Future commits will implement a button to toggle the state between
true and false.

For the moment, **there are no user facing changes** in this commit.