Add expand/collapse all; fix state bug
This PR fixes a bug where a card's expanded state was reset to the
default (collapsed) if the card was filtered out of the search results.
I did this by lifting the `expanded` state up from `<StudentCard />` to
its parent `<StudentList />` and refactoring it to be a list of ids
rather than a boolean.

This refactor also cleared the way to implementing expand / collapse
all buttons.

There are also styling changes so the control bar (input + buttons)
stays in a fixed position while the content (search results) scrolls
past it.

This required a refactor of `<Container />` so that it is scrollable
only when the `scroll` prop is true. To prevent breaking any existing
uses of `<Container />`, the default value of this prop is true.